worker host part 1
