+++ ext/runtime_ops/worker_host.rs
@@ -1,20 +1,15 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-use crate::ops::TestingFeaturesEnabled;
-use crate::permissions::create_child_permissions;
-use crate::permissions::ChildPermissionsArg;
-use crate::permissions::PermissionsContainer;
-use crate::web_worker::run_web_worker;
-use crate::web_worker::SendableWebWorkerHandle;
-use crate::web_worker::WebWorker;
-use crate::web_worker::WebWorkerHandle;
-use crate::web_worker::WebWorkerType;
-use crate::web_worker::WorkerControlEvent;
-use crate::web_worker::WorkerId;
-use crate::worker::FormatJsErrorFn;
 use deno_core::error::AnyError;
+use deno_core::error::JsError;
 use deno_core::op2;
+use deno_core::serde::de;
 use deno_core::serde::Deserialize;
+use deno_core::serde::Deserializer;
+use deno_core::serde::Serialize;
+use deno_core::serde::Serializer;
+use deno_core::serde_json;
+use deno_core::serde_json::json;
 use deno_core::CancelFuture;
 use deno_core::CancelHandle;
 use deno_core::ModuleSpecifier;
@@ -23,353 +18,447 @@ use deno_web::JsMessageData;
 use log::debug;
 use std::cell::RefCell;
 use std::collections::HashMap;
+use std::fmt;
 use std::rc::Rc;
 use std::sync::Arc;
 
+#[derive(
+  Debug, Default, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize,
+)]
+pub struct WorkerId(u32);
+impl fmt::Display for WorkerId {
+  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+    write!(f, "worker-{}", self.0)
+  }
+}
+impl WorkerId {
+  pub fn next(&self) -> Option<WorkerId> {
+    self.0.checked_add(1).map(WorkerId)
+  }
+}
+
+pub type FormatJsErrorFn = dyn Fn(&JsError) -> String + Sync + Send;
+
+/// Events that are sent to host from child
+/// worker.
+pub enum WorkerControlEvent {
+  Error(AnyError),
+  TerminalError(AnyError),
+  Close,
+}
+
+impl Serialize for WorkerControlEvent {
+  fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+  where
+    S: Serializer,
+  {
+    let type_id = match &self {
+      WorkerControlEvent::TerminalError(_) => 1_i32,
+      WorkerControlEvent::Error(_) => 2_i32,
+      WorkerControlEvent::Close => 3_i32,
+    };
+
+    match self {
+      WorkerControlEvent::TerminalError(error)
+      | WorkerControlEvent::Error(error) => {
+        let value = match error.downcast_ref::<JsError>() {
+          Some(js_error) => {
+            let frame = js_error.frames.iter().find(|f| match &f.file_name {
+              Some(s) => !s.trim_start_matches('[').starts_with("ext:"),
+              None => false,
+            });
+            json!({
+              "message": js_error.exception_message,
+              "fileName": frame.map(|f| f.file_name.as_ref()),
+              "lineNumber": frame.map(|f| f.line_number.as_ref()),
+              "columnNumber": frame.map(|f| f.column_number.as_ref()),
+            })
+          }
+          None => json!({
+            "message": error.to_string(),
+          }),
+        };
+
+        Serialize::serialize(&(type_id, value), serializer)
+      }
+      _ => Serialize::serialize(&(type_id, ()), serializer),
+    }
+  }
+}
+
+pub trait WorkerHost {
+  fn create_worker(
+    state: &mut OpState,
+    args: CreateWorkerArgs,
+  ) -> Result<WorkerId, AnyError> {
+    unimplemented!()
+  }
+  fn terminate_worker(state: &mut OpState, id: WorkerId) {
+    unimplemented!()
+  }
+  fn post_message(
+    state: &mut OpState,
+    id: WorkerId,
+    data: JsMessageData,
+  ) -> Result<(), AnyError> {
+    unimplemented!()
+  }
+  fn recv_ctrl(
+    state: Rc<RefCell<OpState>>,
+    id: WorkerId,
+  ) -> Result<WorkerControlEvent, AnyError> {
+    unimplemented!()
+  }
+  fn recv_message(
+    state: Rc<RefCell<OpState>>,
+    id: WorkerId,
+  ) -> Result<Option<JsMessageData>, AnyError> {
+    unimplemented!()
+  }
+}
+
 pub const UNSTABLE_FEATURE_NAME: &str = "worker";
-
-pub struct CreateWebWorkerArgs {
-  pub name: String,
-  pub worker_id: WorkerId,
-  pub parent_permissions: PermissionsContainer,
-  pub permissions: PermissionsContainer,
-  pub main_module: ModuleSpecifier,
-  pub worker_type: WebWorkerType,
-}
-
-pub type CreateWebWorkerCb = dyn Fn(CreateWebWorkerArgs) -> (WebWorker, SendableWebWorkerHandle)
-  + Sync
-  + Send;
-
-/// A holder for callback that is used to create a new
-/// WebWorker. It's a struct instead of a type alias
-/// because `GothamState` used in `OpState` overrides
-/// value if type aliases have the same underlying type
-#[derive(Clone)]
-struct CreateWebWorkerCbHolder(Arc<CreateWebWorkerCb>);
-
-#[derive(Clone)]
-struct FormatJsErrorFnHolder(Option<Arc<FormatJsErrorFn>>);
-
-pub struct WorkerThread {
-  worker_handle: WebWorkerHandle,
-  cancel_handle: Rc<CancelHandle>,
-
-  // A WorkerThread that hasn't been explicitly terminated can only be removed
-  // from the WorkersTable once close messages have been received for both the
-  // control and message channels. See `close_channel`.
-  ctrl_closed: bool,
-  message_closed: bool,
-}
-
-impl WorkerThread {
-  fn terminate(self) {
-    // Cancel recv ops when terminating the worker, so they don't show up as
-    // pending ops.
-    self.cancel_handle.cancel();
-  }
-}
-
-impl Drop for WorkerThread {
-  fn drop(&mut self) {
-    self.worker_handle.clone().terminate();
-  }
-}
-
-pub type WorkersTable = HashMap<WorkerId, WorkerThread>;
 
 deno_core::extension!(
   deno_worker_host,
+  parameters = [P: WorkerHost],
   ops = [
-    op_create_worker,
-    op_host_terminate_worker,
-    op_host_post_message,
-    op_host_recv_ctrl,
-    op_host_recv_message,
+    op_create_worker<P>,
+    op_host_terminate_worker<P>,
+    op_host_post_message<P>,
+    op_host_recv_ctrl<P>,
+    op_host_recv_message<P>,
   ],
-  options = {
-    create_web_worker_cb: Arc<CreateWebWorkerCb>,
-    format_js_error_fn: Option<Arc<FormatJsErrorFn>>,
-  },
-  state = |state, options| {
-    state.put::<WorkersTable>(WorkersTable::default());
+  state = |state| {
     state.put::<WorkerId>(WorkerId::default());
-
-    let create_web_worker_cb_holder =
-      CreateWebWorkerCbHolder(options.create_web_worker_cb);
-    state.put::<CreateWebWorkerCbHolder>(create_web_worker_cb_holder);
-    let format_js_error_fn_holder =
-      FormatJsErrorFnHolder(options.format_js_error_fn);
-    state.put::<FormatJsErrorFnHolder>(format_js_error_fn_holder);
   },
 );
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum WebWorkerType {
+  Classic,
+  Module,
+}
 
 #[derive(Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct CreateWorkerArgs {
-  has_source_code: bool,
-  name: Option<String>,
-  permissions: Option<ChildPermissionsArg>,
-  source_code: String,
-  specifier: String,
-  worker_type: WebWorkerType,
+  pub has_source_code: bool,
+  pub name: Option<String>,
+  pub permissions: Option<ChildPermissionsArg>,
+  pub source_code: String,
+  pub specifier: String,
+  pub worker_type: WebWorkerType,
 }
 
 /// Create worker as the host
 #[op2]
 #[serde]
-fn op_create_worker(
+fn op_create_worker<H>(
   state: &mut OpState,
   #[serde] args: CreateWorkerArgs,
-) -> Result<WorkerId, AnyError> {
-  let specifier = args.specifier.clone();
-  let maybe_source_code = if args.has_source_code {
-    Some(args.source_code.clone())
-  } else {
-    None
-  };
-  let args_name = args.name;
-  let worker_type = args.worker_type;
-  if let WebWorkerType::Classic = worker_type {
-    if let TestingFeaturesEnabled(false) = state.borrow() {
-      return Err(
-        deno_webstorage::DomExceptionNotSupportedError::new(
-          "Classic workers are not supported.",
-        )
-        .into(),
-      );
-    }
-  }
-
-  if args.permissions.is_some() {
-    super::check_unstable(
-      state,
-      UNSTABLE_FEATURE_NAME,
-      "Worker.deno.permissions",
-    );
-  }
-  let parent_permissions = state.borrow_mut::<PermissionsContainer>();
-  let worker_permissions = if let Some(child_permissions_arg) = args.permissions
-  {
-    let mut parent_permissions = parent_permissions.0.lock();
-    let perms =
-      create_child_permissions(&mut parent_permissions, child_permissions_arg)?;
-    PermissionsContainer::new(perms)
-  } else {
-    parent_permissions.clone()
-  };
-  let parent_permissions = parent_permissions.clone();
-  let worker_id = state.take::<WorkerId>();
-  let create_web_worker_cb = state.borrow::<CreateWebWorkerCbHolder>().clone();
-  let format_js_error_fn = state.borrow::<FormatJsErrorFnHolder>().clone();
-  state.put::<WorkerId>(worker_id.next().unwrap());
-
-  let module_specifier = deno_core::resolve_url(&specifier)?;
-  let worker_name = args_name.unwrap_or_default();
-
-  let (handle_sender, handle_receiver) = std::sync::mpsc::sync_channel::<
-    Result<SendableWebWorkerHandle, AnyError>,
-  >(1);
-
-  // Setup new thread
-  let thread_builder = std::thread::Builder::new().name(format!("{worker_id}"));
-
-  // Spawn it
-  thread_builder.spawn(move || {
-    // Any error inside this block is terminal:
-    // - JS worker is useless - meaning it throws an exception and can't do anything else,
-    //  all action done upon it should be noops
-    // - newly spawned thread exits
-
-    let (worker, external_handle) =
-      (create_web_worker_cb.0)(CreateWebWorkerArgs {
-        name: worker_name,
-        worker_id,
-        parent_permissions,
-        permissions: worker_permissions,
-        main_module: module_specifier.clone(),
-        worker_type,
-      });
-
-    // Send thread safe handle from newly created worker to host thread
-    handle_sender.send(Ok(external_handle)).unwrap();
-    drop(handle_sender);
-
-    // At this point the only method of communication with host
-    // is using `worker.internal_channels`.
-    //
-    // Host can already push messages and interact with worker.
-    run_web_worker(
-      worker,
-      module_specifier,
-      maybe_source_code,
-      format_js_error_fn.0,
-    )
-  })?;
-
-  // Receive WebWorkerHandle from newly created worker
-  let worker_handle = handle_receiver.recv().unwrap()?;
-
-  let worker_thread = WorkerThread {
-    worker_handle: worker_handle.into(),
-    cancel_handle: CancelHandle::new_rc(),
-    ctrl_closed: false,
-    message_closed: false,
-  };
-
-  // At this point all interactions with worker happen using thread
-  // safe handler returned from previous function calls
-  state
-    .borrow_mut::<WorkersTable>()
-    .insert(worker_id, worker_thread);
-
-  Ok(worker_id)
+) -> Result<WorkerId, AnyError>
+where
+  H: WorkerHost + 'static,
+{
+  H::create_worker(state, args)
 }
 
 #[op2]
-fn op_host_terminate_worker(state: &mut OpState, #[serde] id: WorkerId) {
-  if let Some(worker_thread) = state.borrow_mut::<WorkersTable>().remove(&id) {
-    worker_thread.terminate();
-  } else {
-    debug!("tried to terminate non-existent worker {}", id);
-  }
-}
-
-enum WorkerChannel {
-  Ctrl,
-  Messages,
-}
-
-/// Close a worker's channel. If this results in both of a worker's channels
-/// being closed, the worker will be removed from the workers table.
-fn close_channel(
-  state: Rc<RefCell<OpState>>,
-  id: WorkerId,
-  channel: WorkerChannel,
-) {
-  use std::collections::hash_map::Entry;
-
-  let mut s = state.borrow_mut();
-  let workers = s.borrow_mut::<WorkersTable>();
-
-  // `Worker.terminate()` might have been called already, meaning that we won't
-  // find the worker in the table - in that case ignore.
-  if let Entry::Occupied(mut entry) = workers.entry(id) {
-    let terminate = {
-      let worker_thread = entry.get_mut();
-      match channel {
-        WorkerChannel::Ctrl => {
-          worker_thread.ctrl_closed = true;
-          worker_thread.message_closed
-        }
-        WorkerChannel::Messages => {
-          worker_thread.message_closed = true;
-          worker_thread.ctrl_closed
-        }
-      }
-    };
-
-    if terminate {
-      entry.remove().terminate();
-    }
-  }
+fn op_host_terminate_worker<H>(state: &mut OpState, #[serde] id: WorkerId)
+where
+  H: WorkerHost + 'static,
+{
+  H::terminate_worker(state, id)
 }
 
 /// Get control event from guest worker as host
 #[op2(async)]
 #[serde]
-async fn op_host_recv_ctrl(
+async fn op_host_recv_ctrl<H>(
   state: Rc<RefCell<OpState>>,
   #[serde] id: WorkerId,
-) -> Result<WorkerControlEvent, AnyError> {
-  let (worker_handle, cancel_handle) = {
-    let state = state.borrow();
-    let workers_table = state.borrow::<WorkersTable>();
-    let maybe_handle = workers_table.get(&id);
-    if let Some(handle) = maybe_handle {
-      (handle.worker_handle.clone(), handle.cancel_handle.clone())
-    } else {
-      // If handle was not found it means worker has already shutdown
-      return Ok(WorkerControlEvent::Close);
-    }
-  };
-
-  let maybe_event = worker_handle
-    .get_control_event()
-    .or_cancel(cancel_handle)
-    .await;
-  match maybe_event {
-    Ok(Ok(Some(event))) => {
-      // Terminal error means that worker should be removed from worker table.
-      if let WorkerControlEvent::TerminalError(_) = &event {
-        close_channel(state, id, WorkerChannel::Ctrl);
-      }
-      Ok(event)
-    }
-    Ok(Ok(None)) => {
-      // If there was no event from worker it means it has already been closed.
-      close_channel(state, id, WorkerChannel::Ctrl);
-      Ok(WorkerControlEvent::Close)
-    }
-    Ok(Err(err)) => Err(err),
-    Err(_) => {
-      // The worker was terminated.
-      Ok(WorkerControlEvent::Close)
-    }
-  }
+) -> Result<WorkerControlEvent, AnyError>
+where
+  H: WorkerHost + 'static,
+{
+  H::recv_ctrl(state, id)
 }
 
 #[op2(async)]
 #[serde]
-async fn op_host_recv_message(
+async fn op_host_recv_message<H>(
   state: Rc<RefCell<OpState>>,
   #[serde] id: WorkerId,
-) -> Result<Option<JsMessageData>, AnyError> {
-  let (worker_handle, cancel_handle) = {
-    let s = state.borrow();
-    let workers_table = s.borrow::<WorkersTable>();
-    let maybe_handle = workers_table.get(&id);
-    if let Some(handle) = maybe_handle {
-      (handle.worker_handle.clone(), handle.cancel_handle.clone())
-    } else {
-      // If handle was not found it means worker has already shutdown
-      return Ok(None);
-    }
-  };
-
-  let ret = worker_handle
-    .port
-    .recv(state.clone())
-    .or_cancel(cancel_handle)
-    .await;
-  match ret {
-    Ok(Ok(ret)) => {
-      if ret.is_none() {
-        close_channel(state, id, WorkerChannel::Messages);
-      }
-      Ok(ret)
-    }
-    Ok(Err(err)) => Err(err),
-    Err(_) => {
-      // The worker was terminated.
-      Ok(None)
-    }
-  }
+) -> Result<Option<JsMessageData>, AnyError>
+where
+  H: WorkerHost + 'static,
+{
+  H::recv_message(state, id)
 }
 
 /// Post message to guest worker as host
 #[op2]
-fn op_host_post_message(
+fn op_host_post_message<H>(
   state: &mut OpState,
   #[serde] id: WorkerId,
   #[serde] data: JsMessageData,
-) -> Result<(), AnyError> {
-  if let Some(worker_thread) = state.borrow::<WorkersTable>().get(&id) {
-    debug!("post message to worker {}", id);
-    let worker_handle = worker_thread.worker_handle.clone();
-    worker_handle.port.send(state, data)?;
-  } else {
-    debug!("tried to post message to non-existent worker {}", id);
+) -> Result<(), AnyError>
+where
+  H: WorkerHost + 'static,
+{
+  H::post_message(state, id, data)
+}
+
+#[derive(Debug, Eq, PartialEq)]
+pub enum ChildUnitPermissionArg {
+  Inherit,
+  Granted,
+  NotGranted,
+}
+
+impl<'de> Deserialize<'de> for ChildUnitPermissionArg {
+  fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+  where
+    D: Deserializer<'de>,
+  {
+    struct ChildUnitPermissionArgVisitor;
+    impl<'de> de::Visitor<'de> for ChildUnitPermissionArgVisitor {
+      type Value = ChildUnitPermissionArg;
+
+      fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+        formatter.write_str("\"inherit\" or boolean")
+      }
+
+      fn visit_unit<E>(self) -> Result<ChildUnitPermissionArg, E>
+      where
+        E: de::Error,
+      {
+        Ok(ChildUnitPermissionArg::NotGranted)
+      }
+
+      fn visit_str<E>(self, v: &str) -> Result<ChildUnitPermissionArg, E>
+      where
+        E: de::Error,
+      {
+        if v == "inherit" {
+          Ok(ChildUnitPermissionArg::Inherit)
+        } else {
+          Err(de::Error::invalid_value(de::Unexpected::Str(v), &self))
+        }
+      }
+
+      fn visit_bool<E>(self, v: bool) -> Result<ChildUnitPermissionArg, E>
+      where
+        E: de::Error,
+      {
+        match v {
+          true => Ok(ChildUnitPermissionArg::Granted),
+          false => Ok(ChildUnitPermissionArg::NotGranted),
+        }
+      }
+    }
+    deserializer.deserialize_any(ChildUnitPermissionArgVisitor)
   }
-  Ok(())
+}
+
+#[derive(Debug, Eq, PartialEq)]
+pub enum ChildUnaryPermissionArg {
+  Inherit,
+  Granted,
+  NotGranted,
+  GrantedList(Vec<String>),
+}
+
+impl<'de> Deserialize<'de> for ChildUnaryPermissionArg {
+  fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+  where
+    D: Deserializer<'de>,
+  {
+    struct ChildUnaryPermissionArgVisitor;
+    impl<'de> de::Visitor<'de> for ChildUnaryPermissionArgVisitor {
+      type Value = ChildUnaryPermissionArg;
+
+      fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+        formatter.write_str("\"inherit\" or boolean or string[]")
+      }
+
+      fn visit_unit<E>(self) -> Result<ChildUnaryPermissionArg, E>
+      where
+        E: de::Error,
+      {
+        Ok(ChildUnaryPermissionArg::NotGranted)
+      }
+
+      fn visit_str<E>(self, v: &str) -> Result<ChildUnaryPermissionArg, E>
+      where
+        E: de::Error,
+      {
+        if v == "inherit" {
+          Ok(ChildUnaryPermissionArg::Inherit)
+        } else {
+          Err(de::Error::invalid_value(de::Unexpected::Str(v), &self))
+        }
+      }
+
+      fn visit_bool<E>(self, v: bool) -> Result<ChildUnaryPermissionArg, E>
+      where
+        E: de::Error,
+      {
+        match v {
+          true => Ok(ChildUnaryPermissionArg::Granted),
+          false => Ok(ChildUnaryPermissionArg::NotGranted),
+        }
+      }
+
+      fn visit_seq<V>(
+        self,
+        mut v: V,
+      ) -> Result<ChildUnaryPermissionArg, V::Error>
+      where
+        V: de::SeqAccess<'de>,
+      {
+        let mut granted_list = vec![];
+        while let Some(value) = v.next_element::<String>()? {
+          granted_list.push(value);
+        }
+        Ok(ChildUnaryPermissionArg::GrantedList(granted_list))
+      }
+    }
+    deserializer.deserialize_any(ChildUnaryPermissionArgVisitor)
+  }
+}
+
+/// Directly deserializable from JS worker and test permission options.
+#[derive(Debug, Eq, PartialEq)]
+pub struct ChildPermissionsArg {
+  pub env: ChildUnaryPermissionArg,
+  pub hrtime: ChildUnitPermissionArg,
+  pub net: ChildUnaryPermissionArg,
+  pub ffi: ChildUnaryPermissionArg,
+  pub read: ChildUnaryPermissionArg,
+  pub run: ChildUnaryPermissionArg,
+  pub sys: ChildUnaryPermissionArg,
+  pub write: ChildUnaryPermissionArg,
+}
+
+impl ChildPermissionsArg {
+  pub fn inherit() -> Self {
+    ChildPermissionsArg {
+      env: ChildUnaryPermissionArg::Inherit,
+      hrtime: ChildUnitPermissionArg::Inherit,
+      net: ChildUnaryPermissionArg::Inherit,
+      ffi: ChildUnaryPermissionArg::Inherit,
+      read: ChildUnaryPermissionArg::Inherit,
+      run: ChildUnaryPermissionArg::Inherit,
+      sys: ChildUnaryPermissionArg::Inherit,
+      write: ChildUnaryPermissionArg::Inherit,
+    }
+  }
+
+  pub fn none() -> Self {
+    ChildPermissionsArg {
+      env: ChildUnaryPermissionArg::NotGranted,
+      hrtime: ChildUnitPermissionArg::NotGranted,
+      net: ChildUnaryPermissionArg::NotGranted,
+      ffi: ChildUnaryPermissionArg::NotGranted,
+      read: ChildUnaryPermissionArg::NotGranted,
+      run: ChildUnaryPermissionArg::NotGranted,
+      sys: ChildUnaryPermissionArg::NotGranted,
+      write: ChildUnaryPermissionArg::NotGranted,
+    }
+  }
+}
+
+impl<'de> Deserialize<'de> for ChildPermissionsArg {
+  fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+  where
+    D: Deserializer<'de>,
+  {
+    struct ChildPermissionsArgVisitor;
+    impl<'de> de::Visitor<'de> for ChildPermissionsArgVisitor {
+      type Value = ChildPermissionsArg;
+
+      fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+        formatter.write_str("\"inherit\" or \"none\" or object")
+      }
+
+      fn visit_unit<E>(self) -> Result<ChildPermissionsArg, E>
+      where
+        E: de::Error,
+      {
+        Ok(ChildPermissionsArg::inherit())
+      }
+
+      fn visit_str<E>(self, v: &str) -> Result<ChildPermissionsArg, E>
+      where
+        E: de::Error,
+      {
+        if v == "inherit" {
+          Ok(ChildPermissionsArg::inherit())
+        } else if v == "none" {
+          Ok(ChildPermissionsArg::none())
+        } else {
+          Err(de::Error::invalid_value(de::Unexpected::Str(v), &self))
+        }
+      }
+
+      fn visit_map<V>(self, mut v: V) -> Result<ChildPermissionsArg, V::Error>
+      where
+        V: de::MapAccess<'de>,
+      {
+        let mut child_permissions_arg = ChildPermissionsArg::none();
+        while let Some((key, value)) =
+          v.next_entry::<String, serde_json::Value>()?
+        {
+          if key == "env" {
+            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
+            child_permissions_arg.env = arg.map_err(|e| {
+              de::Error::custom(format!("(deno.permissions.env) {e}"))
+            })?;
+          } else if key == "hrtime" {
+            let arg = serde_json::from_value::<ChildUnitPermissionArg>(value);
+            child_permissions_arg.hrtime = arg.map_err(|e| {
+              de::Error::custom(format!("(deno.permissions.hrtime) {e}"))
+            })?;
+          } else if key == "net" {
+            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
+            child_permissions_arg.net = arg.map_err(|e| {
+              de::Error::custom(format!("(deno.permissions.net) {e}"))
+            })?;
+          } else if key == "ffi" {
+            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
+            child_permissions_arg.ffi = arg.map_err(|e| {
+              de::Error::custom(format!("(deno.permissions.ffi) {e}"))
+            })?;
+          } else if key == "read" {
+            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
+            child_permissions_arg.read = arg.map_err(|e| {
+              de::Error::custom(format!("(deno.permissions.read) {e}"))
+            })?;
+          } else if key == "run" {
+            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
+            child_permissions_arg.run = arg.map_err(|e| {
+              de::Error::custom(format!("(deno.permissions.run) {e}"))
+            })?;
+          } else if key == "sys" {
+            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
+            child_permissions_arg.sys = arg.map_err(|e| {
+              de::Error::custom(format!("(deno.permissions.sys) {e}"))
+            })?;
+          } else if key == "write" {
+            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
+            child_permissions_arg.write = arg.map_err(|e| {
+              de::Error::custom(format!("(deno.permissions.write) {e}"))
+            })?;
+          } else {
+            return Err(de::Error::custom("unknown permission name"));
+          }
+        }
+        Ok(child_permissions_arg)
+      }
+    }
+    deserializer.deserialize_any(ChildPermissionsArgVisitor)
+  }
 }
