+++ runtime/lib.rs
@@ -14,13 +14,13 @@ pub use deno_kv;
 pub use deno_napi;
 pub use deno_net;
 pub use deno_node;
+pub use deno_runtime_ops;
 pub use deno_tls;
 pub use deno_url;
 pub use deno_web;
 pub use deno_webidl;
 pub use deno_websocket;
 pub use deno_webstorage;
-pub use deno_runtime_ops;
 
 pub mod colors;
 pub mod errors;
