+++ runtime/permissions/mod.rs
@@ -15,6 +15,9 @@ use deno_core::serde_json;
 use deno_core::url;
 use deno_core::url::Url;
 use deno_core::ModuleSpecifier;
+pub use deno_runtime_ops::worker_host::ChildPermissionsArg;
+use deno_runtime_ops::worker_host::ChildUnaryPermissionArg;
+use deno_runtime_ops::worker_host::ChildUnitPermissionArg;
 use log;
 use once_cell::sync::Lazy;
 use std::borrow::Cow;
@@ -27,7 +30,6 @@ use std::str::FromStr;
 use std::string::ToString;
 use std::sync::Arc;
 use which::which;
-
 mod prompter;
 use prompter::permission_prompt;
 use prompter::PromptResponse;
@@ -1660,258 +1662,6 @@ fn escalation_error() -> AnyError {
     "PermissionDenied",
     "Can't escalate parent thread permissions",
   )
-}
-
-#[derive(Debug, Eq, PartialEq)]
-pub enum ChildUnitPermissionArg {
-  Inherit,
-  Granted,
-  NotGranted,
-}
-
-impl<'de> Deserialize<'de> for ChildUnitPermissionArg {
-  fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-  where
-    D: Deserializer<'de>,
-  {
-    struct ChildUnitPermissionArgVisitor;
-    impl<'de> de::Visitor<'de> for ChildUnitPermissionArgVisitor {
-      type Value = ChildUnitPermissionArg;
-
-      fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        formatter.write_str("\"inherit\" or boolean")
-      }
-
-      fn visit_unit<E>(self) -> Result<ChildUnitPermissionArg, E>
-      where
-        E: de::Error,
-      {
-        Ok(ChildUnitPermissionArg::NotGranted)
-      }
-
-      fn visit_str<E>(self, v: &str) -> Result<ChildUnitPermissionArg, E>
-      where
-        E: de::Error,
-      {
-        if v == "inherit" {
-          Ok(ChildUnitPermissionArg::Inherit)
-        } else {
-          Err(de::Error::invalid_value(de::Unexpected::Str(v), &self))
-        }
-      }
-
-      fn visit_bool<E>(self, v: bool) -> Result<ChildUnitPermissionArg, E>
-      where
-        E: de::Error,
-      {
-        match v {
-          true => Ok(ChildUnitPermissionArg::Granted),
-          false => Ok(ChildUnitPermissionArg::NotGranted),
-        }
-      }
-    }
-    deserializer.deserialize_any(ChildUnitPermissionArgVisitor)
-  }
-}
-
-#[derive(Debug, Eq, PartialEq)]
-pub enum ChildUnaryPermissionArg {
-  Inherit,
-  Granted,
-  NotGranted,
-  GrantedList(Vec<String>),
-}
-
-impl<'de> Deserialize<'de> for ChildUnaryPermissionArg {
-  fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-  where
-    D: Deserializer<'de>,
-  {
-    struct ChildUnaryPermissionArgVisitor;
-    impl<'de> de::Visitor<'de> for ChildUnaryPermissionArgVisitor {
-      type Value = ChildUnaryPermissionArg;
-
-      fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        formatter.write_str("\"inherit\" or boolean or string[]")
-      }
-
-      fn visit_unit<E>(self) -> Result<ChildUnaryPermissionArg, E>
-      where
-        E: de::Error,
-      {
-        Ok(ChildUnaryPermissionArg::NotGranted)
-      }
-
-      fn visit_str<E>(self, v: &str) -> Result<ChildUnaryPermissionArg, E>
-      where
-        E: de::Error,
-      {
-        if v == "inherit" {
-          Ok(ChildUnaryPermissionArg::Inherit)
-        } else {
-          Err(de::Error::invalid_value(de::Unexpected::Str(v), &self))
-        }
-      }
-
-      fn visit_bool<E>(self, v: bool) -> Result<ChildUnaryPermissionArg, E>
-      where
-        E: de::Error,
-      {
-        match v {
-          true => Ok(ChildUnaryPermissionArg::Granted),
-          false => Ok(ChildUnaryPermissionArg::NotGranted),
-        }
-      }
-
-      fn visit_seq<V>(
-        self,
-        mut v: V,
-      ) -> Result<ChildUnaryPermissionArg, V::Error>
-      where
-        V: de::SeqAccess<'de>,
-      {
-        let mut granted_list = vec![];
-        while let Some(value) = v.next_element::<String>()? {
-          granted_list.push(value);
-        }
-        Ok(ChildUnaryPermissionArg::GrantedList(granted_list))
-      }
-    }
-    deserializer.deserialize_any(ChildUnaryPermissionArgVisitor)
-  }
-}
-
-/// Directly deserializable from JS worker and test permission options.
-#[derive(Debug, Eq, PartialEq)]
-pub struct ChildPermissionsArg {
-  env: ChildUnaryPermissionArg,
-  hrtime: ChildUnitPermissionArg,
-  net: ChildUnaryPermissionArg,
-  ffi: ChildUnaryPermissionArg,
-  read: ChildUnaryPermissionArg,
-  run: ChildUnaryPermissionArg,
-  sys: ChildUnaryPermissionArg,
-  write: ChildUnaryPermissionArg,
-}
-
-impl ChildPermissionsArg {
-  pub fn inherit() -> Self {
-    ChildPermissionsArg {
-      env: ChildUnaryPermissionArg::Inherit,
-      hrtime: ChildUnitPermissionArg::Inherit,
-      net: ChildUnaryPermissionArg::Inherit,
-      ffi: ChildUnaryPermissionArg::Inherit,
-      read: ChildUnaryPermissionArg::Inherit,
-      run: ChildUnaryPermissionArg::Inherit,
-      sys: ChildUnaryPermissionArg::Inherit,
-      write: ChildUnaryPermissionArg::Inherit,
-    }
-  }
-
-  pub fn none() -> Self {
-    ChildPermissionsArg {
-      env: ChildUnaryPermissionArg::NotGranted,
-      hrtime: ChildUnitPermissionArg::NotGranted,
-      net: ChildUnaryPermissionArg::NotGranted,
-      ffi: ChildUnaryPermissionArg::NotGranted,
-      read: ChildUnaryPermissionArg::NotGranted,
-      run: ChildUnaryPermissionArg::NotGranted,
-      sys: ChildUnaryPermissionArg::NotGranted,
-      write: ChildUnaryPermissionArg::NotGranted,
-    }
-  }
-}
-
-impl<'de> Deserialize<'de> for ChildPermissionsArg {
-  fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-  where
-    D: Deserializer<'de>,
-  {
-    struct ChildPermissionsArgVisitor;
-    impl<'de> de::Visitor<'de> for ChildPermissionsArgVisitor {
-      type Value = ChildPermissionsArg;
-
-      fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        formatter.write_str("\"inherit\" or \"none\" or object")
-      }
-
-      fn visit_unit<E>(self) -> Result<ChildPermissionsArg, E>
-      where
-        E: de::Error,
-      {
-        Ok(ChildPermissionsArg::inherit())
-      }
-
-      fn visit_str<E>(self, v: &str) -> Result<ChildPermissionsArg, E>
-      where
-        E: de::Error,
-      {
-        if v == "inherit" {
-          Ok(ChildPermissionsArg::inherit())
-        } else if v == "none" {
-          Ok(ChildPermissionsArg::none())
-        } else {
-          Err(de::Error::invalid_value(de::Unexpected::Str(v), &self))
-        }
-      }
-
-      fn visit_map<V>(self, mut v: V) -> Result<ChildPermissionsArg, V::Error>
-      where
-        V: de::MapAccess<'de>,
-      {
-        let mut child_permissions_arg = ChildPermissionsArg::none();
-        while let Some((key, value)) =
-          v.next_entry::<String, serde_json::Value>()?
-        {
-          if key == "env" {
-            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
-            child_permissions_arg.env = arg.map_err(|e| {
-              de::Error::custom(format!("(deno.permissions.env) {e}"))
-            })?;
-          } else if key == "hrtime" {
-            let arg = serde_json::from_value::<ChildUnitPermissionArg>(value);
-            child_permissions_arg.hrtime = arg.map_err(|e| {
-              de::Error::custom(format!("(deno.permissions.hrtime) {e}"))
-            })?;
-          } else if key == "net" {
-            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
-            child_permissions_arg.net = arg.map_err(|e| {
-              de::Error::custom(format!("(deno.permissions.net) {e}"))
-            })?;
-          } else if key == "ffi" {
-            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
-            child_permissions_arg.ffi = arg.map_err(|e| {
-              de::Error::custom(format!("(deno.permissions.ffi) {e}"))
-            })?;
-          } else if key == "read" {
-            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
-            child_permissions_arg.read = arg.map_err(|e| {
-              de::Error::custom(format!("(deno.permissions.read) {e}"))
-            })?;
-          } else if key == "run" {
-            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
-            child_permissions_arg.run = arg.map_err(|e| {
-              de::Error::custom(format!("(deno.permissions.run) {e}"))
-            })?;
-          } else if key == "sys" {
-            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
-            child_permissions_arg.sys = arg.map_err(|e| {
-              de::Error::custom(format!("(deno.permissions.sys) {e}"))
-            })?;
-          } else if key == "write" {
-            let arg = serde_json::from_value::<ChildUnaryPermissionArg>(value);
-            child_permissions_arg.write = arg.map_err(|e| {
-              de::Error::custom(format!("(deno.permissions.write) {e}"))
-            })?;
-          } else {
-            return Err(de::Error::custom("unknown permission name"));
-          }
-        }
-        Ok(child_permissions_arg)
-      }
-    }
-    deserializer.deserialize_any(ChildPermissionsArgVisitor)
-  }
 }
 
 pub fn create_child_permissions(
